Rename test data to sample data
In order to not use `test` in names outside of testing we rename test
data preparation into sample data preparation. To increase readibility,
we move the respective code into its own folder.
